+++ src/app/feedback/evaluate/_components/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useEvaluateRadio } from './useEvaluateRadio';
