+++ src/constants/index.ts
@@ -22,3 +22,4 @@ export {
 export { REMIND_TIME_TEXT } from '@/constants/remindTimeText';
 export { INPUT_MAX_LENGTH } from '@/constants/userInputMaxLength';
 export { address } from '@/constants/addressRegex';
+export { options } from '@/constants/radio';
