+++ src/constants/radio.ts
@@ -1,0 +1,9 @@
+import { Option } from '@/types';
+
+export const options: Option[] = [
+  { value: '100', label: '매우 잘함 (100%)' },
+  { value: '75', label: '잘 함 (75%)' },
+  { value: '50', label: '보통 (50%)' },
+  { value: '25', label: '대체로 못함 (25%)' },
+  { value: '0', label: '전혀 못함 (0%)' },
+];
