+++ src/types/Radio.ts
@@ -1,0 +1,9 @@
+export type EvaluateRadioProps = {
+  evaluateOption: number;
+  setEvaluateOption: React.Dispatch<React.SetStateAction<number>>;
+};
+
+export type Option = {
+  value: string;
+  label: string;
+};
